% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{add_numbers}
\alias{add_numbers}
\title{Add numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{a number to be added.}

\item{y}{another number}
}
\value{

}
\description{
Add numbers
}
\examples{
add_numbers(3, 4)
}
