% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{multiply_three}
\alias{multiply_three}
\title{Multiply by three
Multiplies a given number by three}
\usage{
multiply_three(n)
}
\arguments{
\item{n}{}
}
\value{
number
}
\description{
Multiply by three
Multiplies a given number by three
}
\examples{
multiply_three(4)
}
